Create Chi_squared_analysis.sql
In order to determine whether the city can meaningfully affect a cities star rating,
we will apply a Chi-Squared analysis of the data. This is a statistical test to det-
ermine whether the stars associated with a city are random or if they are related to
one another. The chi-squared value is given by 

chi_squared=SUM(Observed_i-Expected_i)^2/Expected_i

for each city. The expected value is just the average star rating for each city and
the observed is each individual star rating for that city. 

Each city will have an associatd degree of freedom to it. This is given by

degree of freedom=COUNT(city)-1.

These degrees of freedom are essentially how many different businesses are in a city.
Given a signifcance level of 0.10, the degrees of freedom determine the critical val-
ue for the chi_squared distribution. If the calculated chi_squared value is less than
the critical value for the given city (hence given degrees of freedom), we say that 
we are 90% confident that the city does not influence the star rating for a business.
If it is greater than that critical value, we say that we are 90% confident that the 
city does influence the star rating for a business.

From here we choose values have a number of ratings between 50 and 200. We pick gre-
ater than 50 because we need a larger sample size to justify the test and less than
200 since most tables don't have calculated critical values for degrees of freedom
(num_trials_per_city) greater than 200. 

When we apply this chi_squared analysis, we find that every city is below the 
corresponding critical value. This tells us that the given test is not statistically
meaningful in answering the problem at hand! 

As an example, the critical value given Richmond Hill is 66.548. This value is much 
larger than the calculated chi_squared meaning we can't determine with any signifcance 
through this test whether the city influences the star rating of the busineeses.